Fix the problem of dependent but prohibited fields.
In the case when the field is prohibited for the role, but it is in
forced_depends_on list, the field will be removed from the sql query, so
another one field which is dependant on the first one will not be able
to be computed.
The problem described above is resolved in the commit.